+++ crates/dxm-manifest/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `artifact::TXADMIN_MONITOR` constant.
 - Optional third-party monitor in artifact and lockfile definitions.
 - Profile structures.
+- Resource patches.
 
 ### Changed